requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2

SQLAlchemy==2.0.32
psycopg[binary]==3.2.1